@@ -1,15 +1,8 @@
-[nosetests]
-with-coverage=1
-cover-package=kivy
-with-id=1
-verbosity=2
-logging-level=DEBUG
-
 [kivy]
 cython_min=0.24
-cython_max=0.29.26
+cython_max=0.29.32
 cython_exclude=0.27,0.27.2
-python_versions=3.6 - 3.10
+python_versions=3.7 - 3.10
 
 [coverage:run]
 parallel = True
@@ -21,14 +14,14 @@ plugins =
 concurrency = thread, multiprocessing
 
 [options]
-python_requires = >=3.6
+python_requires = >=3.7
 install_requires =
     Kivy-Garden>=0.1.4
     docutils
     pygments
-    kivy_deps.angle~=0.3.1; sys_platform == "win32"
-    kivy_deps.sdl2~=0.4.4; sys_platform == "win32"
-    kivy_deps.glew~=0.3.0; sys_platform == "win32"
+    kivy_deps.angle~=0.3.2; sys_platform == "win32"
+    kivy_deps.sdl2~=0.5.0; sys_platform == "win32"
+    kivy_deps.glew~=0.3.1; sys_platform == "win32"
     pypiwin32; sys_platform == "win32"
 dependency_links = https://github.com/kivy-garden/garden/archive/master.zip
 
@@ -40,7 +33,7 @@ dev =
     pytest_asyncio!=0.11.0
     pytest-timeout
     pytest-benchmark
-    PyInstaller @ https://github.com/pyinstaller/pyinstaller/archive/develop.zip
+    pyinstaller
     sphinx
     sphinxcontrib-blockdiag
     sphinxcontrib-seqdiag
@@ -48,17 +41,19 @@ dev =
     sphinxcontrib-nwdiag
     funcparserlib==1.0.0a0
     kivy_deps.gstreamer_dev~=0.3.3; sys_platform == "win32"
-    kivy_deps.sdl2_dev~=0.4.4; sys_platform == "win32"
-    kivy_deps.glew_dev~=0.3.0; sys_platform == "win32"
+    kivy_deps.sdl2_dev~=0.5.0; sys_platform == "win32"
+    kivy_deps.glew_dev~=0.3.1; sys_platform == "win32"
     flake8
     pre-commit
+    responses
 base =
     pillow
+    requests
     docutils
     pygments
-    kivy_deps.angle~=0.3.1; sys_platform == "win32"
-    kivy_deps.sdl2~=0.4.4; sys_platform == "win32"
-    kivy_deps.glew~=0.3.0; sys_platform == "win32"
+    kivy_deps.angle~=0.3.2; sys_platform == "win32"
+    kivy_deps.sdl2~=0.5.0; sys_platform == "win32"
+    kivy_deps.glew~=0.3.1; sys_platform == "win32"
     pypiwin32; sys_platform == "win32"
 media =
     kivy_deps.gstreamer~=0.3.3; sys_platform == "win32"
@@ -68,24 +63,25 @@ full =
     docutils
     pygments
     kivy_deps.gstreamer~=0.3.3; sys_platform == "win32"
-    kivy_deps.angle~=0.3.1; sys_platform == "win32"
-    kivy_deps.sdl2~=0.4.4; sys_platform == "win32"
-    kivy_deps.glew~=0.3.0; sys_platform == "win32"
+    kivy_deps.angle~=0.3.2; sys_platform == "win32"
+    kivy_deps.sdl2~=0.5.0; sys_platform == "win32"
+    kivy_deps.glew~=0.3.1; sys_platform == "win32"
     ffpyplayer; sys_platform == "linux" or sys_platform == "darwin"
     pypiwin32; sys_platform == "win32"
 gstreamer =
     kivy_deps.gstreamer~=0.3.3; sys_platform == "win32"
+    # don't use 0.4.0 because it was deleted
 angle =
-    kivy_deps.angle~=0.3.1; sys_platform == "win32"
+    kivy_deps.angle~=0.3.2; sys_platform == "win32"
 sdl2 =
-    kivy_deps.sdl2~=0.4.4; sys_platform == "win32"
+    kivy_deps.sdl2~=0.5.0; sys_platform == "win32"
 glew =
-    kivy_deps.glew~=0.3.0; sys_platform == "win32"
+    kivy_deps.glew~=0.3.1; sys_platform == "win32"
 
 [flake8]
 ignore = E125,E126,E127,E128,E402,E741,E731,W503,F401,W504,F841,E722
 max-line-length = 80
-exclude = __pycache__,.tox,.git/,doc/,build/,.eggs/
+exclude = __pycache__,.tox,.git/,doc/,build/,.eggs/,venv/
 statistics = true
 show-source = true
 count = true
