@@ -93,14 +93,18 @@ class MacMotionEvent(MotionEvent):
     and shape profiles.
     '''
 
+    def __init__(self, *args, **kwargs):
+        kwargs.setdefault('is_touch', True)
+        kwargs.setdefault('type_id', 'touch')
+        super().__init__(*args, **kwargs)
+        self.profile = ('pos', 'shape')
+
     def depack(self, args):
-        self.is_touch = True
         self.shape = ShapeRect()
         self.sx, self.sy = args[0], args[1]
         self.shape.width = args[2]
         self.shape.height = args[2]
-        self.profile = ('pos', 'shape')
-        super(MacMotionEvent, self).depack(args)
+        super().depack(args)
 
     def __str__(self):
         return '<MacMotionEvent id=%d pos=(%f, %f) device=%s>' \
