@@ -31,10 +31,12 @@ __all__ = (
 import sys
 import shutil
 from getopt import getopt, GetoptError
+import os
 from os import environ, mkdir
 from os.path import dirname, join, basename, exists, expanduser
 import pkgutil
 import re
+import importlib
 from kivy.logger import Logger, LOG_LEVELS
 from kivy.utils import platform
 from kivy._version import __version__, RELEASE as _KIVY_RELEASE, \
@@ -162,7 +164,10 @@ def kivy_register_post_configuration(callback):
 
 
 def kivy_usage():
-    '''Kivy Usage: %s [OPTION...]::
+    '''Kivy Usage: %s [KIVY OPTION...] [-- PROGRAM OPTIONS]::
+
+            Options placed after a '-- ' separator, will not be touched by kivy,
+            and instead passed to your program.
 
             Set KIVY_NO_ARGS=1 in your environment or before you import Kivy to
             disable Kivy's argument parser.
@@ -265,6 +270,22 @@ for examples_dir in (
         kivy_examples_dir = examples_dir
         break
 
+
+def _patch_mod_deps_win(dep_mod, mod_name):
+    import site
+    dep_bins = []
+
+    for d in [sys.prefix, site.USER_BASE]:
+        p = join(d, 'share', mod_name, 'bin')
+        if os.path.isdir(p):
+            os.environ["PATH"] = p + os.pathsep + os.environ["PATH"]
+            if hasattr(os, 'add_dll_directory'):
+                os.add_dll_directory(p)
+            dep_bins.append(p)
+
+    dep_mod.dep_bins = dep_bins
+
+
 # if there are deps, import them so they can do their magic.
 _packages = []
 try:
@@ -294,7 +315,9 @@ except ImportError:
 _logging_msgs = []
 for importer, modname, package in _packages:
     try:
-        mod = importer.find_module(modname).load_module(modname)
+        module_spec = importer.find_spec(modname)
+        mod = importlib.util.module_from_spec(module_spec)
+        module_spec.loader.exec_module(mod)
 
         version = ''
         if hasattr(mod, '__version__'):
@@ -302,6 +325,10 @@ for importer, modname, package in _packages:
         _logging_msgs.append(
             'deps: Successfully imported "{}.{}"{}'.
             format(package, modname, version))
+
+        if modname.startswith('gst') and version == '0.3.3':
+            _patch_mod_deps_win(mod, modname)
+
     except ImportError as e:
         Logger.warning(
             'deps: Error importing dependency "{}.{}": {}'.
@@ -314,7 +341,7 @@ if any(name in sys.argv[0] for name in (
     environ['KIVY_DOC'] = '1'
 if 'sphinx-build' in sys.argv[0]:
     environ['KIVY_DOC_INCLUDE'] = '1'
-if any(('nosetests' in arg or 'pytest' in arg) for arg in sys.argv):
+if any('pytest' in arg for arg in sys.argv):
     environ['KIVY_UNITTEST'] = '1'
 if any('pyinstaller' in arg.lower() for arg in sys.argv):
     environ['KIVY_PACKAGING'] = '1'
