@@ -1,7 +1,7 @@
 # This file is imported from __init__.py and exec'd from setup.py
 
 MAJOR = 2
-MINOR = 1
+MINOR = 2
 MICRO = 0
 RELEASE = False
 
