@@ -9,33 +9,71 @@ pointing devices (touch and non-touch). This class defines all the properties
 and methods needed to handle 2D and 3D movements but has many more
 capabilities.
 
-.. note::
+Usually you would never need to create the :class:`MotionEvent` yourself as
+this is the role of the :mod:`~kivy.input.providers`.
 
-    You never create the :class:`MotionEvent` yourself: this is the role of the
-    :mod:`~kivy.input.providers`.
+Flow of the motion events
+-------------------------
 
-Motion Event and Touch
-----------------------
+1. The :class:`MotionEvent` 's are gathered from input providers by
+   :class:`~kivy.base.EventLoopBase`.
+2. Post processing is performed by registered processors
+   :mod:`~kivy.input.postproc`.
+3. :class:`~kivy.base.EventLoopBase` dispatches all motion events using
+   `on_motion` event to all registered listeners including the
+   :class:`~kivy.core.window.WindowBase`.
+4. Once received in :meth:`~kivy.core.window.WindowBase.on_motion` events
+   (touch or non-touch) are all registered managers. If a touch event is not
+   handled by at least one manager, then it is dispatched through
+   :meth:`~kivy.core.window.WindowBase.on_touch_down`,
+   :meth:`~kivy.core.window.WindowBase.on_touch_move` and
+   :meth:`~kivy.core.window.WindowBase.on_touch_up`.
+5. Widgets receive events in :meth:`~kivy.uix.widget.Widget.on_motion` method
+   (if passed by a manager) or on `on_touch_xxx` methods.
 
-We differentiate between a Motion Event and Touch event. A Touch event is a
-:class:`MotionEvent` with the `pos` profile. Only these events are dispatched
-throughout the widget tree.
+Motion events and event managers
+--------------------------------
 
-1. The :class:`MotionEvent` 's are gathered from input providers.
-2. All the :class:`MotionEvent` 's are dispatched from
-    :meth:`~kivy.core.window.WindowBase.on_motion`.
-3. If a :class:`MotionEvent` has a `pos` profile, we dispatch it through
-    :meth:`~kivy.core.window.WindowBase.on_touch_down`,
-    :meth:`~kivy.core.window.WindowBase.on_touch_move` and
-    :meth:`~kivy.core.window.WindowBase.on_touch_up`.
+A motion event is a touch event if its :attr:`MotionEvent.is_touch` is set to
+`True`. Beside `is_touch` attribute, :attr:`MotionEvent.type_id` can be used to
+check for event's general type. Currently two types are dispatched by
+input providers: "touch" and "hover".
 
-Listening to a Motion Event
+Event managers can be used to dispatch any motion event throughout the widget
+tree and a manager uses `type_id` to specify which event types it want to
+receive. See :mod:`~kivy.eventmanager` to learn how to define and register
+an event manager.
+
+A manager can also assign a new `type_id` to
+:attr:`MotionEvent.type_id` before dispatching it to the widgets. This useful
+when dispatching a specific event::
+
+    class MouseTouchManager(EventManagerBase):
+
+        type_ids = ('touch',)
+
+        def dispatch(self, etype, me):
+            accepted = False
+            if me.device == 'mouse':
+                me.push() # Save current type_id and other values
+                me.type_id = 'mouse_touch'
+                self.window.transform_motion_event_2d(me)
+                # Dispatch mouse touch event to widgets which registered
+                # to receive 'mouse_touch'
+                for widget in self.window.children[:]:
+                    if widget.dispatch('on_motion', etype, me):
+                        accepted = True
+                        break
+                me.pop() # Restore
+            return accepted
+
+Listening to a motion event
 ---------------------------
 
-If you want to receive all MotionEvents, Touch or not, you can bind the
+If you want to receive all motion events, touch or not, you can bind the
 MotionEvent from the :class:`~kivy.core.window.Window` to your own callback::
 
-    def on_motion(self, etype, motionevent):
+    def on_motion(self, etype, me):
         # will receive all motion events.
         pass
 
@@ -90,6 +128,8 @@ import weakref
 from inspect import isroutine
 from copy import copy
 from time import time
+
+from kivy.eventmanager import MODE_DEFAULT_DISPATCH
 from kivy.vector import Vector
 
 
@@ -123,7 +163,7 @@ MotionEventBase = MotionEventMetaclass('MotionEvent', (object, ), {})
 
 
 class MotionEvent(MotionEventBase):
-    '''Abstract class that represents an input event (touch or non-touch).
+    '''Abstract class that represents an input event.
 
     :Parameters:
         `id`: str
@@ -135,7 +175,7 @@ class MotionEvent(MotionEventBase):
     __uniq_id = 0
     __attrs__ = \
         ('device', 'push_attrs', 'push_attrs_stack',
-         'is_touch', 'id', 'shape', 'profile',
+         'is_touch', 'type_id', 'id', 'dispatch_mode', 'shape', 'profile',
          # current position, in 0-1 range
          'sx', 'sy', 'sz',
          # first position set, in 0-1 range
@@ -157,25 +197,38 @@ class MotionEvent(MotionEventBase):
          'is_triple_tap', 'triple_tap_time',
          'ud')
 
-    def __init__(self, device, id, args, is_touch=False):
+    def __init__(self, device, id, args, is_touch=False, type_id=None):
         if self.__class__ == MotionEvent:
             raise NotImplementedError('class MotionEvent is abstract')
         MotionEvent.__uniq_id += 1
 
-        #: True if the Motion Event is a Touch.
+        #: True if the MotionEvent is a touch.
         self.is_touch = is_touch
+
+        #: (Experimental) String to identify event type.
+        #:
+        #: .. versionadded:: 2.1.0
+        self.type_id = type_id
+
+        #: (Experimental) Used by a event manager or a widget to assign
+        #: the dispatching mode. Defaults to
+        #: :const:`~kivy.eventmanager.MODE_DEFAULT_DISPATCH`. See
+        #: :mod:`~kivy.eventmanager` for available modes.
+        #:
+        #: .. versionadded:: 2.1.0
+        self.dispatch_mode = MODE_DEFAULT_DISPATCH
 
         #: Attributes to push by default, when we use :meth:`push` : x, y, z,
         #: dx, dy, dz, ox, oy, oz, px, py, pz.
         self.push_attrs_stack = []
         self.push_attrs = ('x', 'y', 'z', 'dx', 'dy', 'dz', 'ox', 'oy', 'oz',
-                           'px', 'py', 'pz', 'pos')
+                           'px', 'py', 'pz', 'pos', 'type_id', 'dispatch_mode')
 
-        #: Uniq ID of the touch. You can safely use this property, it will be
-        #: never the same across all existing touches.
+        #: Uniq ID of the event. You can safely use this property, it will be
+        #: never the same across all existing events.
         self.uid = MotionEvent.__uniq_id
 
-        #: Device used for creating this touch
+        #: Device used for creating this event.
         self.device = device
 
         # For grab
@@ -183,32 +236,32 @@ class MotionEvent(MotionEventBase):
         self.grab_exclusive_class = None
         self.grab_state = False
 
-        #: Used to determine which widget the touch is being dispatched to.
+        #: Used to determine which widget the event is being dispatched to.
         #: Check the :meth:`grab` function for more information.
         self.grab_current = None
 
-        #: Currently pressed button
+        #: Currently pressed button.
         self.button = None
 
-        #: Profiles currently used in the touch
+        #: Profiles currently used in the event.
         self.profile = []
 
-        #: Id of the touch, not uniq. This is generally the Id set by the input
-        #: provider, like ID in TUIO. If you have multiple TUIO source,
-        #: the same id can be used. Prefer to use :attr:`uid` attribute
+        #: Id of the event, not unique. This is generally the Id set by the
+        #: input provider, like ID in TUIO. If you have multiple TUIO sources,
+        #: then same id can be used. Prefer to use :attr:`uid` attribute
         #: instead.
         self.id = id
 
-        #: Shape of the touch, subclass of
+        #: Shape of the touch event, subclass of
         #: :class:`~kivy.input.shape.Shape`.
-        #: By default, the property is set to None
+        #: By default, the property is set to None.
         self.shape = None
 
-        #: X position, in 0-1 range
+        #: X position, in 0-1 range.
         self.sx = 0.0
-        #: Y position, in 0-1 range
+        #: Y position, in 0-1 range.
         self.sy = 0.0
-        #: Z position, in 0-1 range
+        #: Z position, in 0-1 range.
         self.sz = 0.0
         #: Origin X position, in 0-1 range.
         self.osx = None
@@ -228,46 +281,46 @@ class MotionEvent(MotionEventBase):
         self.dsy = None
         #: Delta between self.sz and self.psz, in 0-1 range.
         self.dsz = None
-        #: X position, in window range
+        #: X position, in window range.
         self.x = 0.0
-        #: Y position, in window range
+        #: Y position, in window range.
         self.y = 0.0
-        #: Z position, in window range
+        #: Z position, in window range.
         self.z = 0.0
-        #: Origin X position, in window range
+        #: Origin X position, in window range.
         self.ox = None
-        #: Origin Y position, in window range
+        #: Origin Y position, in window range.
         self.oy = None
-        #: Origin Z position, in window range
+        #: Origin Z position, in window range.
         self.oz = None
-        #: Previous X position, in window range
+        #: Previous X position, in window range.
         self.px = None
-        #: Previous Y position, in window range
+        #: Previous Y position, in window range.
         self.py = None
-        #: Previous Z position, in window range
+        #: Previous Z position, in window range.
         self.pz = None
-        #: Delta between self.x and self.px, in window range
+        #: Delta between self.x and self.px, in window range.
         self.dx = None
-        #: Delta between self.y and self.py, in window range
+        #: Delta between self.y and self.py, in window range.
         self.dy = None
-        #: Delta between self.z and self.pz, in window range
+        #: Delta between self.z and self.pz, in window range.
         self.dz = None
-        #: Position (X, Y), in window range
+        #: Position (X, Y), in window range.
         self.pos = (0.0, 0.0)
 
-        #: Initial time of the touch creation
+        #: Initial time of the event creation.
         self.time_start = time()
 
-        #: Time of the last update
+        #: Time of the last update.
         self.time_update = self.time_start
 
-        #: Time of the end event (last touch usage)
+        #: Time of the end event (last event usage).
         self.time_end = -1
 
-        #: Indicate if the touch is a double tap or not
+        #: Indicate if the touch event is a double tap or not.
         self.is_double_tap = False
 
-        #: Indicate if the touch is a triple tap or not
+        #: Indicate if the touch event is a triple tap or not.
         #:
         #: .. versionadded:: 1.7.0
         self.is_triple_tap = False
@@ -283,7 +336,7 @@ class MotionEvent(MotionEventBase):
         self.triple_tap_time = 0
 
         #: User data dictionary. Use this dictionary to save your own data on
-        #: the touch.
+        #: the event.
         self.ud = EnhancedDictionary()
 
         #: If set to `True` (default) keeps first previous position
@@ -322,10 +375,12 @@ class MotionEvent(MotionEventBase):
         self.dsz = self.sz - self.psz
 
     def grab(self, class_instance, exclusive=False):
-        '''Grab this motion event. You can grab a touch if you want
-        to receive subsequent :meth:`~kivy.uix.widget.Widget.on_touch_move`
-        and :meth:`~kivy.uix.widget.Widget.on_touch_up`
-        events, even if the touch is not dispatched by the parent:
+        '''Grab this motion event.
+
+        If this event is a touch you can grab it if you want to receive
+        subsequent :meth:`~kivy.uix.widget.Widget.on_touch_move` and
+        :meth:`~kivy.uix.widget.Widget.on_touch_up` events, even if the touch
+        is not dispatched by the parent:
 
         .. code-block:: python
 
@@ -350,14 +405,14 @@ class MotionEvent(MotionEventBase):
             Allowed grab for non-touch events.
         '''
         if self.grab_exclusive_class is not None:
-            raise Exception('Cannot grab the touch, touch is exclusive')
+            raise Exception('Event is exclusive and cannot be grabbed')
         class_instance = weakref.ref(class_instance.__self__)
         if exclusive:
             self.grab_exclusive_class = class_instance
         self.grab_list.append(class_instance)
 
     def ungrab(self, class_instance):
-        '''Ungrab a previously grabbed touch
+        '''Ungrab a previously grabbed motion event.
         '''
         class_instance = weakref.ref(class_instance.__self__)
         if self.grab_exclusive_class == class_instance:
@@ -376,7 +431,7 @@ class MotionEvent(MotionEventBase):
         self._first_dispatch_done = True
 
     def move(self, args):
-        '''Move the touch to another position.
+        '''Move to another position.
         '''
         if self.sync_with_dispatch:
             if self._keep_prev_pos:
@@ -406,7 +461,7 @@ class MotionEvent(MotionEventBase):
         self.pz = self.psz * z_max
         if smode:
             # Adjust y for keyboard height
-            if smode == 'pan':
+            if smode == 'pan' or smode == 'below_target':
                 self.y -= kheight
                 self.oy -= kheight
                 self.py -= kheight
@@ -443,7 +498,7 @@ class MotionEvent(MotionEventBase):
                          'valid values are 0, 90, 180 or 270' % rotation)
 
     def push(self, attrs=None):
-        '''Push attribute values in `attrs` onto the stack
+        '''Push attribute values in `attrs` onto the stack.
         '''
         if attrs is None:
             attrs = self.push_attrs
@@ -451,7 +506,7 @@ class MotionEvent(MotionEventBase):
         self.push_attrs_stack.append((attrs, values))
 
     def pop(self):
-        '''Pop attributes values from the stack
+        '''Pop attributes values from the stack.
         '''
         attrs, values = self.push_attrs_stack.pop()
         for i in range(len(attrs)):
@@ -459,7 +514,7 @@ class MotionEvent(MotionEventBase):
 
     def apply_transform_2d(self, transform):
         '''Apply a transformation on x, y, z, px, py, pz,
-        ox, oy, oz, dx, dy, dz
+        ox, oy, oz, dx, dy, dz.
         '''
         self.x, self.y = self.pos = transform(self.x, self.y)
         self.px, self.py = transform(self.px, self.py)
@@ -468,12 +523,12 @@ class MotionEvent(MotionEventBase):
         self.dy = self.y - self.py
 
     def copy_to(self, to):
-        '''Copy some attribute to another touch object.'''
+        '''Copy some attribute to another motion event object.'''
         for attr in self.__attrs__:
             to.__setattr__(attr, copy(self.__getattribute__(attr)))
 
     def distance(self, other_touch):
-        '''Return the distance between the current touch and another touch.
+        '''Return the distance between the two events.
         '''
         return Vector(self.pos).distance(other_touch.pos)
 
@@ -484,25 +539,25 @@ class MotionEvent(MotionEventBase):
     @property
     def dpos(self):
         '''Return delta between last position and current position, in the
-        screen coordinate system (self.dx, self.dy)'''
+        screen coordinate system (self.dx, self.dy).'''
         return self.dx, self.dy
 
     @property
     def opos(self):
-        '''Return the initial position of the touch in the screen
-        coordinate system (self.ox, self.oy)'''
+        '''Return the initial position of the motion event in the screen
+        coordinate system (self.ox, self.oy).'''
         return self.ox, self.oy
 
     @property
     def ppos(self):
-        '''Return the previous position of the touch in the screen
-        coordinate system (self.px, self.py)'''
+        '''Return the previous position of the motion event in the screen
+        coordinate system (self.px, self.py).'''
         return self.px, self.py
 
     @property
     def spos(self):
-        '''Return the position in the 0-1 coordinate system
-        (self.sx, self.sy)'''
+        '''Return the position in the 0-1 coordinate system (self.sx, self.sy).
+        '''
         return self.sx, self.sy
 
     def __str__(self):
@@ -525,7 +580,7 @@ class MotionEvent(MotionEventBase):
 
     @property
     def is_mouse_scrolling(self, *args):
-        '''Returns True if the touch is a mousewheel scrolling
+        '''Returns True if the touch event is a mousewheel scrolling
 
         .. versionadded:: 1.6.0
         '''
