@@ -38,7 +38,7 @@ _delayed_start = None
 
 
 class BuilderException(ParserException):
-    '''Exception raised when the Builder failed to apply a rule on a widget.
+    '''Exception raised when the Builder fails to apply a rule on a widget.
     '''
     pass
 
@@ -85,7 +85,7 @@ def update_intermediates(base, keys, bound, s, fn, args, instance, value):
     ''' Function that is called when an intermediate property is updated
     and `rebind` of that property is True. In that case, we unbind
     all bound funcs that were bound to attrs of the old value of the
-    property and rebind to the new value of the property.
+    property and rebind them to the new value of the property.
 
     For example, if the rule is `self.a.b.c.d`, then when b is changed, we
     unbind from `b`, `c` and `d`, if they were bound before (they were not
@@ -555,13 +555,13 @@ class BuilderBase(object):
         # rootrule: the current root rule (for children of a rule)
 
         # will collect reference to all the id in children
-        assert(rule not in self.rulectx)
+        assert rule not in self.rulectx
         self.rulectx[rule] = rctx = {
             'ids': {'root': widget.proxy_ref},
             'set': [], 'hdl': []}
 
         # extract the context of the rootrule (not rule!)
-        assert(rootrule in self.rulectx)
+        assert rootrule in self.rulectx
         rctx = self.rulectx[rootrule]
 
         # if a template context is passed, put it as "ctx"
@@ -687,7 +687,7 @@ class BuilderBase(object):
             rule = None
             for widget_set, rules in reversed(rctx['set']):
                 for rule in rules:
-                    assert(isinstance(rule, ParserRuleProperty))
+                    assert isinstance(rule, ParserRuleProperty)
                     key = rule.name
                     value = rule.co_value
                     if type(value) is CodeType:
@@ -716,8 +716,8 @@ class BuilderBase(object):
             crule = None
             for widget_set, rules in rctx['hdl']:
                 for crule in rules:
-                    assert(isinstance(crule, ParserRuleProperty))
-                    assert(crule.name.startswith('on_'))
+                    assert isinstance(crule, ParserRuleProperty)
+                    assert crule.name.startswith('on_')
                     key = crule.name
                     if not widget_set.is_event_type(key):
                         key = key[3:]
