@@ -161,7 +161,7 @@ class ButtonBehavior(object):
     def on_touch_up(self, touch):
         if touch.grab_current is not self:
             return super(ButtonBehavior, self).on_touch_up(touch)
-        assert(self in touch.ud)
+        assert self in touch.ud
         touch.ungrab(self)
         self.last_touch = touch
 
