@@ -255,17 +255,17 @@ class CompoundSelectionBehavior(object):
     '''
 
     text_entry_timeout = NumericProperty(1.)
-    '''When typing characters in rapid succession (i.e. the time difference since
-    the last character is less than :attr:`text_entry_timeout`), the keys get
-    concatenated and the combined text is passed as the key argument of
-    :meth:`goto_node`.
+    '''When typing characters in rapid succession (i.e. the time difference
+    since the last character is less than :attr:`text_entry_timeout`), the
+    keys get concatenated and the combined text is passed as the key argument
+    of :meth:`goto_node`.
 
     .. versionadded:: 1.10.0
     '''
 
     _anchor = None  # the last anchor node selected (e.g. shift relative node)
     # the idx may be out of sync
-    _anchor_idx = 0  # cache indexs in case list hasn't changed
+    _anchor_idx = 0  # cache indexes in case list hasn't changed
     _last_selected_node = None  # the absolute last node selected
     _last_node_idx = 0
     _ctrl_down = False  # if it's pressed - for e.g. shift selection
