@@ -289,7 +289,7 @@ class TouchRippleButtonBehavior(TouchRippleBehavior):
     def on_touch_up(self, touch):
         if touch.grab_current is not self:
             return super(TouchRippleButtonBehavior, self).on_touch_up(touch)
-        assert(self in touch.ud)
+        assert self in touch.ud
         touch.ungrab(self)
         self.last_touch = touch
         if self.disabled:
