@@ -669,6 +669,15 @@ class ScrollView(StencilView):
         touch.pop()
         return ret
 
+    def on_motion(self, etype, me):
+        if me.type_id in self.motion_filter and 'pos' in me.profile:
+            me.push()
+            me.apply_transform_2d(self.to_local)
+            ret = super().on_motion(etype, me)
+            me.pop()
+            return ret
+        return super().on_motion(etype, me)
+
     def on_touch_down(self, touch):
         if self.dispatch('on_scroll_start', touch):
             self._touch = touch
@@ -924,7 +933,7 @@ class ScrollView(StencilView):
             if not touch.ud['sv.handled']['y'] and self.do_scroll_y \
                     and self.effect_y:
                 height = self.height
-                if touch.ud.get('in_bar_y', False):
+                if touch.ud.get('in_bar_y', False) and self.vbar[1] != 1.0:
                     dy = touch.dy / float(height - height * self.vbar[1])
                     self.scroll_y = min(max(self.scroll_y + dy, 0.), 1.)
                     self._trigger_update_from_scroll()
