@@ -372,7 +372,7 @@ class TextInputCutCopyPaste(Bubble):
         mode = self.mode
 
         if parent:
-            self.clear_widgets()
+            self.content.clear_widgets()
             if mode == 'paste':
                 # show only paste on long touch
                 self.but_selectall.opacity = 1
@@ -387,7 +387,7 @@ class TextInputCutCopyPaste(Bubble):
                 widget_list = (self.but_cut, self.but_copy, self.but_paste)
 
             for widget in widget_list:
-                self.add_widget(widget)
+                self.content.add_widget(widget)
 
     def do(self, action):
         textinput = self.textinput
@@ -526,6 +526,7 @@ class TextInput(FocusBehavior, Widget):
         self._scroll_distance_x = 0
         self._scroll_distance_y = 0
         self._enable_scroll = True
+        self._have_scrolled = False
 
         # [from; to) range of lines being partially or fully rendered
         # in TextInput's viewport
@@ -1385,8 +1386,10 @@ class TextInput(FocusBehavior, Widget):
             ):
                 cursor_x = i
                 break
+        else:
+            cursor_x = len(lines[cursor_y])
 
-        return int(cursor_x), int(cursor_y)
+        return cursor_x, cursor_y
 
     #
     # Selection control
@@ -1587,6 +1590,12 @@ class TextInput(FocusBehavior, Widget):
         # stores the touch for later use
         self._touch_down = touch
 
+        # Is a new touch_down, so previous scroll states needs to be reset
+        self._enable_scroll = True
+        self._have_scrolled = False
+        self._scroll_distance_x = 0
+        self._scroll_distance_y = 0
+
         self._hide_cut_copy_paste(EventLoop.window)
         # schedule long touch for paste
         self._long_touch_ev = Clock.schedule_once(self.long_touch, .5)
@@ -1621,10 +1630,8 @@ class TextInput(FocusBehavior, Widget):
 
         if self.scroll_from_swipe:
             self.scroll_text_from_swipe(touch)
-        else:
-            self._enable_scroll = False
 
-        if not self._enable_scroll and self._selection_touch is touch:
+        if not self._have_scrolled and self._selection_touch is touch:
             self.cursor = self.get_cursor_from_xy(touch.x, touch.y)
             self._selection_to = self.cursor_index()
             self._update_selection()
@@ -1649,29 +1656,10 @@ class TextInput(FocusBehavior, Widget):
             or self._touch_count == 4
         )
 
-        _scroll_timeout = touch.time_update - touch.time_start
-        # conditions for discarding touch such as swipe to scroll and as
-        # selection on hold. If the conditions are true, the tap will be
-        # generically recognized as a single tap.
-        single_tap = (
-            _scroll_timeout <= self.scroll_timeout / 1000
-            and self._scroll_distance_x <= self.scroll_distance
-            and self._scroll_distance_y <= self.scroll_distance
-        )
-
-        # if the given conditions are true, the touch will be recognized as a
-        # single tap. If there is selected text, the selection will be canceled
-        # and it will be possible to start a new selection.
-        if (
-            self.scroll_from_swipe
-            and not prioritized_touch_types
-            and single_tap
-        ):
+        if not self._have_scrolled and not prioritized_touch_types:
+            # Is a single tap and did not scrolled.
+            # Selection needs to be canceled.
             self._cancel_update_selection(self._touch_down)
-
-        self._enable_scroll = True
-        self._scroll_distance_x = 0
-        self._scroll_distance_y = 0
 
         # show Bubble
         win = EventLoop.window
@@ -1700,52 +1688,51 @@ class TextInput(FocusBehavior, Widget):
             return True
 
     def scroll_text_from_swipe(self, touch):
+        _scroll_timeout = (touch.time_update - touch.time_start) * 1000
+        self._scroll_distance_x += abs(touch.dx)
+        self._scroll_distance_y += abs(touch.dy)
+        if not self._have_scrolled:
+            # To be considered a scroll, touch should travel more than
+            # scroll_distance in less than the scroll_timeout since touch_down
+            if not (
+                _scroll_timeout <= self.scroll_timeout
+                and (
+                    (self._scroll_distance_x >= self.scroll_distance)
+                    or (self._scroll_distance_y >= self.scroll_distance)
+                )
+            ):
+                # Distance isn't enough (yet) to consider it as a scroll
+                if _scroll_timeout <= self.scroll_timeout:
+                    # Timeout is not reached, scroll is still enabled.
+                    return False
+                else:
+                    self._enable_scroll = False
+                    self._cancel_update_selection(self._touch_down)
+                    return False
+            # We have a scroll!
+            self._have_scrolled = True
+
+        self.cancel_long_touch_event()
+
         if self.multiline:
-            _scroll_timeout = touch.time_update - touch.time_start
-            self._scroll_distance_x += abs(touch.dx)
-            self._scroll_distance_y += abs(touch.dy)
-
-            # disable scroll and start selection mode if scroll distance
-            # isn't reached within scroll_timeout
-            if (
-                _scroll_timeout >= self.scroll_timeout / 1000
-                and self._scroll_distance_x <= self.scroll_distance
-                and self._scroll_distance_y <= self.scroll_distance
-            ):
-                self._enable_scroll = False
-                self._cancel_update_selection(self._touch_down)
-
-            if self._enable_scroll:
-                self.cancel_long_touch_event()
-                max_scroll_y = max(0, self.minimum_height - self.height)
-                self.scroll_y = min(max(0, self.scroll_y + touch.dy),
-                                    max_scroll_y)
-                self._trigger_update_graphics()
-                self._position_handles()
-                return True
-
+            max_scroll_y = max(0, self.minimum_height - self.height)
+            self.scroll_y = min(
+                max(0, self.scroll_y + touch.dy),
+                max_scroll_y
+            )
         else:
-            _scroll_timeout = touch.time_update - touch.time_start
-            self._scroll_distance_x += abs(touch.dx)
-
-            # works with the same logic as multiline above
-            if (
-                _scroll_timeout >= self.scroll_timeout / 1000
-                and self._scroll_distance_x <= self.scroll_distance
-            ):
-                self._enable_scroll = False
-                self._cancel_update_selection(self._touch_down)
-
-            if self._enable_scroll:
-                self.cancel_long_touch_event()
-                minimum_width = (self._get_row_width(0) + self.padding[0] +
-                                 self.padding[2])
-                max_scroll_x = max(0, minimum_width - self.width)
-                self.scroll_x = min(max(0, self.scroll_x - touch.dx),
-                                    max_scroll_x)
-                self._trigger_update_graphics()
-                self._position_handles()
-                return True
+            minimum_width = (
+                self._get_row_width(0)
+                + self.padding[0] + self.padding[2]
+            )
+            max_scroll_x = max(0, minimum_width - self.width)
+            self.scroll_x = min(
+                max(0, self.scroll_x - touch.dx),
+                max_scroll_x
+            )
+        self._trigger_update_graphics()
+        self._position_handles()
+        return True
 
     def _handle_pressed(self, instance):
         self._hide_cut_copy_paste()
@@ -2116,7 +2103,7 @@ class TextInput(FocusBehavior, Widget):
 
     def _delete_line(self, idx):
         """Delete current line, and fix cursor position"""
-        assert(idx < len(self._lines))
+        assert idx < len(self._lines)
         self._lines_flags.pop(idx)
         self._lines_labels.pop(idx)
         self._lines.pop(idx)
@@ -2695,11 +2682,13 @@ class TextInput(FocusBehavior, Widget):
             Cache_append('textinput.label', cid, texture)
         return texture
 
+    _tokenize_delimiters = u' .,:;!?\r\t'
+
     def _tokenize(self, text):
         # Tokenize a text string from some delimiters
         if text is None:
             return
-        delimiters = u' .,:;!?\'"\r\t'
+        delimiters = self._tokenize_delimiters
         old_index = 0
         prev_char = ''
         for index, char in enumerate(text):
@@ -2745,9 +2734,14 @@ class TextInput(FocusBehavior, Widget):
         _tab_width, _label_cached = self.tab_width, self._label_cached
 
         # try to add each word on current line.
+        words_widths = {}
         for word in self._tokenize(text):
             is_newline = (word == u'\n')
-            w = text_width(word, _tab_width, _label_cached)
+            try:
+                w = words_widths[word]
+            except KeyError:
+                w = text_width(word, _tab_width, _label_cached)
+                words_widths[word] = w
             # if we have more than the width, or if it's a newline,
             # push the current line, and create a new one
             if (x + w > width and line) or is_newline:
@@ -2763,9 +2757,11 @@ class TextInput(FocusBehavior, Widget):
                     split_width = split_pos = 0
                     # split the word
                     for c in word:
-                        cw = self._get_text_width(
-                            c, self.tab_width, self._label_cached
-                        )
+                        try:
+                            cw = words_widths[c]
+                        except KeyError:
+                            cw = text_width(c, _tab_width, _label_cached)
+                            words_widths[c] = cw
                         if split_width + cw > width:
                             break
                         split_width += cw
@@ -3668,7 +3664,7 @@ class TextInput(FocusBehavior, Widget):
     you can load the system fonts by specifying a font context starting
     with the special string `system://`. This will load the system
     fontconfig configuration, and add your application-specific fonts on
-    top of it (this imposes a signifficant risk of family name collision,
+    top of it (this imposes a significant risk of family name collision,
     Pango may not use your custom font file, but pick one from the system)
 
     .. note::
